Move pydantic model to separate file
